requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==1.3.0
Django==4.2.7
django-filter==23.3
django-import-export==4.3.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.2.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0
